+++ src/views/DesktopView.tsx
@@ -1,0 +1,144 @@
+import { Box, Collapse, SxProps, Theme } from '@mui/material';
+import ShaderPane from '../components/shader/ShaderPane';
+import { type ShaderPaneHandle } from '../components/shader/ShaderPane';
+import { useCallback, useEffect, useState } from 'react';
+
+interface DesktopViewProps {
+	outerContainerRef: React.RefObject<HTMLDivElement>
+	shaderRef: React.RefObject<ShaderPaneHandle>
+	tabBar: React.ReactNode
+	editorContent: React.ReactNode
+	overwriteDialog: React.ReactNode
+	shaderSource: string
+	webcamStream: MediaStream | null
+	audioStream: MediaStream | null
+	webcamEnabled: boolean
+	micEnabled: boolean
+	handleToggleWebcam: () => void
+	handleToggleMic: () => void
+	handleVolumeChange: (value: number) => void
+	handleToggleMute: () => void
+	setShaderError: (error: string | null) => void
+	handleToggleImmersive: () => void
+	immersiveOpacity: number
+	setImmersiveOpacity: (opacity: number) => void
+	editorCollapsed: boolean
+	setEditorCollapsed: (collapsed: boolean) => void
+}
+
+export const DesktopView = ({
+	outerContainerRef,
+	shaderRef,
+	tabBar,
+	editorContent,
+	overwriteDialog,
+	shaderSource,
+	webcamStream,
+	audioStream,
+	webcamEnabled,
+	micEnabled,
+	handleToggleWebcam,
+	handleToggleMic,
+	handleVolumeChange,
+	handleToggleMute,
+	setShaderError,
+	handleToggleImmersive,
+	immersiveOpacity,
+	setImmersiveOpacity,
+	editorCollapsed,
+	setEditorCollapsed,
+}: DesktopViewProps) => {
+
+  const [leftRatio, setLeftRatio] = useState(50)
+
+  const handleHorizontalDividerMouseDown = useCallback((e: React.MouseEvent) => {
+    e.preventDefault()
+    const container = outerContainerRef.current
+    if (!container) return
+    const startX = e.clientX
+    const startRatio = leftRatio
+    const containerW = container.getBoundingClientRect().width
+    const onMove = (me: MouseEvent) => {
+      const delta = me.clientX - startX
+      const newRatio = Math.min(80, Math.max(20, startRatio + (delta / containerW) * 100))
+      setLeftRatio(newRatio)
+    }
+    const onUp = () => {
+      document.removeEventListener('mousemove', onMove)
+      document.removeEventListener('mouseup', onUp)
+    }
+    document.addEventListener('mousemove', onMove)
+    document.addEventListener('mouseup', onUp)
+  }, [leftRatio])
+
+  const collapseSx = {
+    flex: !editorCollapsed ? 1 : undefined,
+    minWidth: 0,
+    display: !editorCollapsed ? 'flex' : undefined,
+    flexDirection: 'column',
+    '& .MuiCollapse-wrapper, & .MuiCollapse-wrapperInner': {
+      width: '100%',
+      height: '100%',
+      display: 'flex',
+      flexDirection: 'column',
+    },
+  } as SxProps<Theme>
+
+
+  useEffect(() => {
+      delete document.documentElement.dataset.immersive
+      document.documentElement.style.removeProperty('--pg-immersive-alpha')
+  }, [immersiveOpacity])
+
+return (
+			<Box ref={outerContainerRef} sx={{ display: 'flex', height: '100vh', width: '100vw', overflow: 'hidden', bgcolor: 'background.app' }}>
+				{/* Left: shader canvas */}
+				<Box sx={{ width: editorCollapsed ? '100%' : `${leftRatio}%`, minWidth: 0, flexShrink: 0 }}>
+					<ShaderPane
+						ref={shaderRef}
+						shaderSource={shaderSource}
+						webcamStream={webcamStream}
+						audioStream={audioStream}
+						webcamEnabled={webcamEnabled}
+						micEnabled={micEnabled}
+						onToggleWebcam={handleToggleWebcam}
+						onToggleMic={handleToggleMic}
+						onVolumeChange={handleVolumeChange}
+						onToggleMute={handleToggleMute}
+						onShaderError={setShaderError}
+						editorCollapsed={editorCollapsed}
+						onToggleEditorCollapsed={() => setEditorCollapsed(!editorCollapsed)}
+						isMobile={false}
+						isImmersive={false}
+						onToggleImmersive={handleToggleImmersive}
+						immersiveOpacity={immersiveOpacity}
+						onImmersiveOpacityChange={setImmersiveOpacity}
+					/>
+				</Box>
+
+				{/* Horizontal drag divider between shader and editor */}
+				{!editorCollapsed && (
+					<Box
+						onMouseDown={handleHorizontalDividerMouseDown}
+						sx={{
+							width: '4px',
+							cursor: 'col-resize',
+							bgcolor: 'border.default',
+							flexShrink: 0,
+							'&:hover': { bgcolor: 'border.faint'},
+						}}
+					/>
+				)}
+
+				{/* Right: editor panel */}
+				<Collapse orientation="horizontal" in={!editorCollapsed} sx={collapseSx}>
+					<Box sx={{ flex: 1, height: '100%', minWidth: 0, display: 'flex', flexDirection: 'column' }}>
+						{tabBar}
+						{editorContent}
+					</Box>
+				</Collapse>
+
+				{overwriteDialog}
+			</Box>
+)
+}
